.history/requirements_20240922154717.txt
amqp
anyio
argon2-cffi
asgiref
attrs
beautifulsoup4
certifi
cffi
charset-normalizer
Django
django-celery-beat
django-celery-results
django-environ
djangorestframework
gunicorn
idna
psycopg2
redis
requests
urllib3
whitenoise
faker
gunicorn
pymysql
dj-database-url
cryptography
boto3 
gspread
Pillow
pandas 
celery
redis 
openpyxl
mysqlclient 
django-allauth